+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.1.0 - 2024-08-17
 - Added ability to remove chests from Mules, Donkeys and Llamas by using shears.
+- Horses can now be Quick Hitched when clicked on an existing leash knot. Previously the knot would've been removed. 
 
 ## 1.0.2 - 2024-08-14
 - Removed leftover debug message in log